@@ -26,12 +26,12 @@ import io.opentelemetry.sdk.OpenTelemetrySdk;
 import io.opentelemetry.sdk.resources.Resource;
 import io.opentelemetry.sdk.trace.SdkTracerProvider;
 import io.opentelemetry.sdk.trace.export.BatchSpanProcessor;
-import io.opentelemetry.semconv.ResourceAttributes;
+import io.opentelemetry.semconv.ServiceAttributes;
 
 public class OpenTelemetrySetup {
     public static OpenTelemetrySdk setup(String serviceName) {
         final Resource resource = Resource.getDefault()
-            .merge(Resource.create(Attributes.of(ResourceAttributes.SERVICE_NAME, serviceName)));
+            .merge(Resource.create(Attributes.of(ServiceAttributes.SERVICE_NAME, serviceName)));
 
         final SdkTracerProvider sdkTracerProvider = SdkTracerProvider.builder()
             .addSpanProcessor(BatchSpanProcessor.builder(LoggingSpanExporter.create()).build())
