@@ -22,8 +22,8 @@ package demo.jaxrs.tracing.server;
 import demo.jaxrs.tracing.OpenTelemetrySetup;
 import io.opentelemetry.sdk.OpenTelemetrySdk;
 import org.apache.cxf.jaxrs.servlet.CXFNonSpringJaxrsServlet;
-import org.eclipse.jetty.ee10.servlet.ServletContextHandler;
-import org.eclipse.jetty.ee10.servlet.ServletHolder;
+import org.eclipse.jetty.ee11.servlet.ServletContextHandler;
+import org.eclipse.jetty.ee11.servlet.ServletHolder;
 
 public class Server {
     protected Server() throws Exception {
