@@ -255,7 +255,7 @@ public abstract class AbstractOpenApiServiceDescriptionTest extends AbstractClie
         try (Response response = uiClient.get()) {
             String html = response.readEntity(String.class);
             assertThat(html, containsString("<!-- HTML"));
-            assertThat(response.getMediaType(), equalTo(MediaType.TEXT_HTML_TYPE));
+            assertThat(response.getMediaType(), equalTo(MediaType.TEXT_HTML_TYPE.withCharset("utf-8")));
         }
     }
 
