@@ -87,7 +87,7 @@ public class SwaggerUiConfigurationQueryConfigTest extends AbstractClientServerT
         try (Response response = uiClient.get()) {
             String html = response.readEntity(String.class);
             assertThat(html, containsString("<!-- HTML"));
-            assertThat(response.getMediaType(), equalTo(MediaType.TEXT_HTML_TYPE));
+            assertThat(response.getMediaType(), equalTo(MediaType.TEXT_HTML_TYPE.withCharset("utf-8")));
         }
     }
 
