@@ -18,6 +18,8 @@
  */
 package org.apache.cxf.systest.jaxrs.sse;
 
+import java.io.IOException;
+import java.io.UncheckedIOException;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.concurrent.CompletableFuture;
@@ -90,6 +92,8 @@ public class BookStore extends BookStoreClientCloseable {
                     sink.close();
                 } catch (final InterruptedException ex) {
                     LOG.error("Communication error", ex);
+                } catch (final IOException ex) {
+                    throw new UncheckedIOException(ex);
                 }
             }
         }.start();
@@ -118,6 +122,8 @@ public class BookStore extends BookStoreClientCloseable {
                     sink.close();
                 } catch (final InterruptedException ex) {
                     LOG.error("Communication error", ex);
+                } catch (final IOException ex) {
+                    throw new UncheckedIOException(ex);
                 }
             }
         }.start();
@@ -137,7 +143,13 @@ public class BookStore extends BookStoreClientCloseable {
                 sink.send(createEvent(builder.name("book"), 4));
                 sink.send(createEvent(builder.name("book"), 5));
             })
-            .whenComplete((r, ex) -> sink.close());
+            .whenComplete((r, ex) -> { 
+                try {
+                    sink.close();
+                } catch (final IOException e) {
+                    throw new UncheckedIOException(e);
+                }
+            });
     }
 
     @GET
@@ -154,13 +166,19 @@ public class BookStore extends BookStoreClientCloseable {
                 sink.send(createRawEvent(builder.name("book"), 4));
                 sink.send(createRawEvent(builder.name("book"), 5));
             })
-            .whenComplete((r, ex) -> sink.close());
+            .whenComplete((r, ex) -> {
+                try {
+                    sink.close();
+                } catch (final IOException e) {
+                    throw new UncheckedIOException(e);
+                }
+            });
     }
 
     @GET
     @Path("nodata")
     @Produces(MediaType.SERVER_SENT_EVENTS)
-    public void nodata(@Context SseEventSink sink) {
+    public void nodata(@Context SseEventSink sink) throws IOException {
         sink.close();
     }
 
@@ -208,6 +226,8 @@ public class BookStore extends BookStoreClientCloseable {
                     sink.close();
                 } catch (final InterruptedException ex) {
                     LOG.error("Communication error", ex);
+                } catch (final IOException ex) {
+                    throw new UncheckedIOException(ex);
                 }
             }
         }.start();
@@ -224,6 +244,8 @@ public class BookStore extends BookStoreClientCloseable {
                     sink.close();
                 } catch (final InterruptedException ex) {
                     LOG.error("Communication error", ex);
+                } catch (final IOException ex) {
+                    throw new UncheckedIOException(ex);
                 }
             }
         }.start();
