@@ -43,12 +43,13 @@ import org.apache.cxf.systest.jaxws.resources.HelloServiceImpl;
 import org.apache.cxf.tracing.micrometer.ObservationClientFeature;
 import org.apache.cxf.tracing.micrometer.ObservationFeature;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.actuate.autoconfigure.observation.web.servlet.WebMvcObservationAutoConfiguration;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.test.autoconfigure.actuate.observability.AutoConfigureObservability;
+import org.springframework.boot.micrometer.metrics.test.autoconfigure.AutoConfigureMetrics;
+import org.springframework.boot.micrometer.tracing.test.autoconfigure.AutoConfigureTracing;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.boot.webmvc.autoconfigure.WebMvcObservationAutoConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.test.context.ActiveProfiles;
 
@@ -71,7 +72,8 @@ import static org.hamcrest.Matchers.hasSize;
         classes = SpringJaxwsObservabilityTest.TestConfig.class
 )
 @ActiveProfiles("jaxws")
-@AutoConfigureObservability
+@AutoConfigureMetrics
+@AutoConfigureTracing
 public class SpringJaxwsObservabilityTest {
 
     private static final String DUMMY_REQUEST_BODY = "<q0:sayHello xmlns:q0=\"http://service.ws.sample/\">"
