@@ -21,7 +21,7 @@ package org.apache.cxf.systest.servlet.resolver;
 
 
 import org.apache.cxf.testutil.common.AbstractBusClientServerTestBase;
-import org.eclipse.jetty.ee10.webapp.WebAppContext;
+import org.eclipse.jetty.ee11.webapp.WebAppContext;
 import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.handler.DefaultHandler;
